requirements.txt
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
ipykernel==6.21.3
ipython==8.11.0
jedi==0.18.2
jupyter_client==8.0.3
jupyter_core==5.2.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
packaging==23.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.1.0
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
pyflakes==3.0.1
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==305
pyzmq==25.0.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
toml==0.10.2
tornado==6.2
traitlets==5.9.0
urllib3==1.26.14
wcwidth==0.2.6
flake8
flake8-docstrings==1.6.0
pytest==6.2.5
python-dotenv==0.19.0
python-telegram-bot==13.7
requests==2.26.0